pyproject.toml
[project]
name = "eytechathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.0",
    "docling>=2.65.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-docling>=2.0.0",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.3",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.5",
    "langgraph-api>=0.5.42",
    "langgraph-cli[inmem]>=0.4.10",
    "langsmith>=0.4.59",
    "markdownify>=1.2.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "tavily-python>=0.7.16",
]
